// packages/yt-dlp/src/type.ts
export interface YtDlpFormat {
  format_id: string;
  format_note: string;
  ext: string;
  protocol: string;
  acodec: string;
  vcodec: string;
  url: string;
  width: number | null;
  height: number | null;
  fragments: any;
  audio_ext: string;
  video_ext: string;
  format: string;
  resolution: string;
  http_headers: any;
  asr: number;
  filesize: number;
  source_preference: number;
  fps: null;
  quality: number;
  tbr: number;
  language: string;
  language_preference: number;
  dynamic_range: null;
  abr: number;
  downloader_options: {
    http_chunk_size: number;
  };
  container: string;
}

export interface YtDlpThumbnail {
  url: string;
  preference: number;
  id: string;
  height?: number;
  width?: number;
  resolution?: string;
}

export interface YtDlpCaption {
  ext: string;
  url: string;
  name: string;
}

export interface YtDlpVideo {
  id: string;
  title: string;
  fulltitle: string;
  url: string;
  ext: string;
  alt_title: string;
  description: string;
  display_id: string;
  uploader: string;
  license: string;
  creator: string;
  timestamp: number;
  upload_date: string;
  release_timestamp: number;
  release_date: string;
  modified_timestamp: number;
  modified_date: string;
  uploader_id: string;
  channel: string;
  channel_id: string;
  channel_follower_count: number;
  location: string;
  duration: number;
  duration_string: string;
  view_count: number;
  like_count: number;
  dislike_count: number;
  repost_count: number;
  average_rating: number;
  comment_count: number;
  age_limit: number;
  live_status: string;
  is_live: boolean;
  was_live: boolean;
  playable_in_embed: string;
  availability: string;
  start_time: number;
  end_time: number;
  format: string;
  format_id: string;
  format_note: string;
  width: number;
  height: number;
  resolution: string;
  tbr: number;
  abr: number;
  acodec: string;
  asr: number;
  vbr: number;
  fps: number;
  dynamic_range: string;
  vcodec: string;
  container: string;
  filesize: number;
  filesize_approx: number;
  protocol: string;
  src: string; // Hack, to be removed
  extractor: string;
  extractor_key: string;
  epoch: number;
  autonumber: number;
  video_autonumber: number;
  n_entries: number;
  playlist_id: string;
  playlist_title: string;
  playlist: string;
  playlist_count: number;
  playlist_index: number;
  playlist_autonumber: number;
  playlist_uploader: string;
  playlist_uploader_id: string;
  webpage_url: string;
  webpage_url_basename: string;
  webpage_url_domain: string;
  original_url: string;
  formats: YtDlpFormat[];
  thumbnails: YtDlpThumbnail[];
  thumbnail: string;
  uploader_url: string;
  channel_url: string;
  categories: string[];
  tags: string[];
  automatic_captions: Record<string, YtDlpCaption[]>;
  subtitles: unknown;
  chapters: any[];
  requested_subtitles: unknown;
  __has_drm: boolean;
  requested_downloads: unknown[];
  source_preference: number;
  quality: number;
  language: string;
  language_preference: number;
  downloader_options: {
    http_chunk_size: number;
  };
  audio_ext: string;
  video_ext: string;
  http_headers: unknown;
  _last_playlist_index: number;
  requested_formats: unknown[];
  stretched_ratio: unknown;
}

export interface YtDlpPlaylist {
  uploader: string;
  uploader_id: string;
  uploader_url: string;
  thumbnails: YtDlpThumbnail[];
  tags: string[];
  view_count: number;
  modified_date: string;
  playlist_count: number;
  channel_follower_count: number | null;
  channel: string;
  channel_id: string;
  channel_url: string;
  id: string;
  title: string;
  description: string;
  entries: YtDlpVideo[];
  original_url: string;
  webpage_url: string;
  webpage_url_basename: string;
  webpage_url_domain: string;
  extractor: string;
  extractor_key: string;
  epoch: number;
}

export type YtDlpResponse = YtDlpVideo | YtDlpPlaylist;

export interface YtDlpOptions {
  update?: boolean;
}